// ui/src/app/modules/billing/interfaces/payments.interfaces.ts
export interface Payments {
  position: number;
  receivedBy: string;
  creator: string;
  paymentType: string;
  referenceNumber: string;
  status: string;
  createdAt: string;
  receiptNumber: string;
  billAmount: number;
  paidAmount: number;
  gepgpaymentDate: string;
  payerNumber: string;
  payerName: string;
  pspName: string;
  accountNumber: string;
}